refactor: refactor app database
