+++ lib/core/local/table/notification_local.dart
@@ -1,0 +1,8 @@
+import 'package:drift/drift.dart';
+
+class NotificationLocal extends Table {
+  TextColumn get id => text()();
+  TextColumn get thumbnail => text()();
+  TextColumn get description => text()();
+  DateTimeColumn get createdAt => dateTime()();
+}
